Removed Magic Numbers and Added Moved Interface to Game Time
