+++ src/model/interfaces/__init__.py
@@ -12,6 +12,7 @@ from .i_item import IItem
 from .i_player import IPlayer
 from .i_tile import ITile
 from .i_game_over import IGameOver
+from .i_time import ITime
 from .i_turn import ITurn
 
 
@@ -24,4 +25,5 @@ __all__ = [
     'ITile',
     'IGameOver',
     'ITurn',
+    'ITime',
 ]
