+++ src/model/interfaces/i_time.py
@@ -1,0 +1,15 @@
+from abc import ABC, abstractmethod
+
+class ITime(ABC):
+    """Abstract Class for building GameTime Classes"""
+    @abstractmethod
+    def get_current_time(self):
+        ...
+
+    @abstractmethod
+    def increase_current_time(self):
+        ...
+
+    @abstractmethod
+    def is_time_valid(self):
+        ...
